REFACTOR: code refactoring
