+++ src/main/java/com/example/getIt/TestController.java
@@ -1,4 +1,4 @@
-package com.example.getIt.controller;
+package com.example.getIt;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
