+++ src/main/java/com/example/getIt/product/entity/ProductEntity.java
@@ -1,4 +1,4 @@
-package com.example.getIt.entity;
+package com.example.getIt.product.entity;
 
 import com.example.getIt.util.BaseEntity;
 import lombok.Builder;
