+++ src/main/java/com/example/getIt/product/entity/WebsiteEntity.java
@@ -1,4 +1,4 @@
-package com.example.getIt.entity;
+package com.example.getIt.product.entity;
 
 import lombok.Getter;
