+++ src/main/java/com/example/getIt/product/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package com.example.getIt.repository;
+package com.example.getIt.product.repository;
 
-import com.example.getIt.DTO.ProductDTO;
-import com.example.getIt.entity.ProductEntity;
+import com.example.getIt.product.DTO.ProductDTO;
+import com.example.getIt.product.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
